requirements.txt
appscript==1.3.0
chardet==5.2.0
docx2pdf==0.1.8
fpdf==1.7.2
lxml==5.3.0
numpy==2.1.3
pandas==2.2.3
pdf2image==1.17.0
pdfkit==1.0.0
pillow==11.0.0
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytz==2024.2
six==1.16.0
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
XlsxWriter==3.2.0
